backend/deepface_service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
facenet-pytorch>=2.5.3
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.1.0
numpy>=1.24.0
python-multipart>=0.0.6
